+++ src/main/java/org/komamitsu/fluency/Fluency.java
@@ -10,6 +10,7 @@ import org.komamitsu.fluency.sender.Sender;
 import org.komamitsu.fluency.sender.TCPSender;
 import org.komamitsu.fluency.sender.heartbeat.TCPHeartbeater;
 import org.komamitsu.fluency.sender.retry.ExponentialBackOffRetryStrategy;
+import org.komamitsu.fluency.util.EventTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -150,8 +151,14 @@ public class Fluency
     public void emit(String tag, long timestamp, Map<String, Object> data)
             throws IOException
     {
+        emit(tag, EventTime.fromTimestamp(timestamp), data);
+    }
+
+    public void emit(String tag, EventTime eventTime, Map<String, Object> data)
+            throws IOException
+    {
         try {
-            buffer.append(tag, timestamp, data);
+            buffer.append(tag, eventTime, data);
             flusher.onUpdate();
         }
         catch (BufferFullException e) {
@@ -164,7 +171,7 @@ public class Fluency
     public void emit(String tag, Map<String, Object> data)
             throws IOException
     {
-        emit(tag, System.currentTimeMillis() / 1000, data);
+        emit(tag, EventTime.fromMillis(System.currentTimeMillis()), data);
     }
 
     @Override