+++ src/test/java/org/komamitsu/fluency/FluencyTest.java
@@ -24,6 +24,7 @@ import org.komamitsu.fluency.sender.failuredetect.FailureDetector;
 import org.komamitsu.fluency.sender.failuredetect.PhiAccrualFailureDetectStrategy;
 import org.komamitsu.fluency.sender.heartbeat.TCPHeartbeater;
 import org.komamitsu.fluency.sender.retry.ExponentialBackOffRetryStrategy;
+import org.komamitsu.fluency.util.EventTime;
 import org.msgpack.value.MapValue;
 import org.msgpack.value.Value;
 import org.slf4j.Logger;
@@ -810,7 +811,7 @@ public class FluencyTest
                 }
 
                 @Override
-                public void onReceive(String tag, long timestampMillis, MapValue data)
+                public void onReceive(String tag, EventTime eventTime, MapValue data)
                 {
                     if (tag.equals("foodb0.bartbl0")) {
                         tag0EventsCounter.incrementAndGet();
@@ -827,7 +828,7 @@ public class FluencyTest
                     else {
                         throw new IllegalArgumentException("Unexpected tag: tag=" + tag);
                     }
-                    assertTrue(startTimestamp <= timestampMillis && timestampMillis < startTimestamp + 60 * 1000);
+                    assertTrue(startTimestamp <= eventTime.getSeconds() && eventTime.getSeconds() < startTimestamp + 60 * 1000);
 
                     assertEquals(4, data.size());
                     for (Map.Entry<Value, Value> kv : data.entrySet()) {
@@ -937,7 +938,7 @@ public class FluencyTest
             Fluency fluency = new Fluency.Builder(stuckSender).setBufferConfig(bufferConfig).build();
             Map<String, Object> event = new HashMap<String, Object>();
             event.put("name", "xxxx");
-            for (int i = 0; i < 7; i++) {
+            for (int i = 0; i < 5; i++) {
                 fluency.emit("tag", event);
             }
             try {